feat: migrate price/amount from i64 to Decimal (NUMERIC)
- resources.price: i64 -> Decimal (in Resource, CreateResourceRequest, UpdateResourceRequest)
- Order/OrderDetail/AdminOrderDetail/PaymentTransaction.amount: i64 -> Decimal
- order_service/payment_service amount params: i64 -> Decimal
- order.rs db: insert use rust_decimal::Decimal after inner doc comment
- handler/order.rs: use ToPrimitive trait for amount.to_f64()
- Add num-traits dependency to rsws_api

+++ rsws_db/src/order.rs
@@ -1,5 +1,6 @@
 //! 订单仓储层
 
+use rust_decimal::Decimal;
 use rsws_common::error::RswsError;
 use rsws_common::error_code::ErrorCode;
 use rsws_common::snowflake;
@@ -22,7 +23,7 @@ impl OrderRepository {
         &self,
         user_id: i64,
         resource_id: i64,
-        amount: i64,
+        amount: Decimal,
         payment_method: &str,
         expire_minutes: i32,
     ) -> Result<Order, RswsError> {
